pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.coverage.report]
show_missing = true

[tool.coverage.run]
branch = true
omit = ["**/__init__.py"]
source = ["uda"]

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 120
multi_line_output = 3
use_parentheses = true

[tool.poetry]
authors = ["Henri Iser <iserhenri@gmail.com>"]
description = "Unsupervised Domain Adaptation"
name = "uda"
version = "0.1.0"

[tool.poetry.dependencies]
jupyter = "^1.0.0"
nibabel = "^3.2.2"
numpy = "^1.22.3"
patchify = "^0.2.3"
python = ">=3.9,<3.11"
pytorch-ignite = "^0.4.9"
scikit-learn = "^1.0.2"
seaborn = "^0.11.2"
seaborn-image = "^0.4.4"
surface-distance-based-measures = {path = "surface-distance"}
torch = "^1.11.0"
tqdm = "^4.64.0"
wandb = "^0.12.16"

[tool.poetry.dev-dependencies]
black = ">=20.8b1"
docformatter = ">=1.3.1"
flake8 = ">=3.8.3"
flake8-annotations = ">=2.1.0"
flake8-bugbear = ">=20.1.4"
flake8-builtins = ">=1.5.3"
flake8-isort = ">=3.0.0"
flake8-use-fstring = ">=1.1"
isort = ">=4.3.21"
nox = ">=2020.5.24"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
pytest-mock = "^3.7.0"
reindent = ">=3.5.1"
toml-sort = "^0.20.0"

[tool.pytest.ini_options]
addopts = "--cov --cov-report=html --cov-report=term"
testpaths = "tests"
